Add more methods for converting Json
